+++ models/stored_genesis_hashes.go
@@ -30,7 +30,7 @@ type StoredGenesisHash struct {
 	CreatedAt      time.Time `json:"created_at" db:"created_at"`
 	UpdatedAt      time.Time `json:"updated_at" db:"updated_at"`
 	GenesisHash    string    `json:"genesisHash" db:"genesis_hash"`
-	FileSizeBytes  int       `json:"fileSizeBytes" db:"file_size_bytes"`
+	FileSizeBytes  uint64    `json:"fileSizeBytes" db:"file_size_bytes"`
 	NumChunks      int       `json:"numChunks" db:"num_chunks"`
 	WebnodeCount   int       `json:"webnodeCount" db:"webnode_count"`
 	Status         int       `json:"status" db:"status"`
